Move Convert to a sub-package
Move Convert to a sub-package to make it possible
to run from other tools as an API
