+++ gocov/convert/convert_test.go
@@ -1,4 +1,4 @@
-package main
+package convert
 
 import (
 	"go/ast"
